Linux build fixes. PostMessage fix.

+++ src/linux/Lgi/GMenuGtk.cpp
@@ -51,7 +51,7 @@ GSubMenu::~GSubMenu()
 	}
 }
 
-int GSubMenu::Length()
+size_t GSubMenu::Length()
 {
 	return Items.Length();
 }
